requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
redis==5.0.1
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
aiocache==0.12.2
tenacity==8.2.3
python-json-logger==2.0.7
prometheus-client==0.19.0
cryptography==42.0.5
pyyaml==6.0.1
